requirements.txt
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
sse-starlette>=2.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
